factor out fusions and utils
