+++ isoquant_lib/fusion/fusion_metadata.py
