+++ isoquant_lib/fusion/fusion_validator.py
