+++ isoquant_tests/test_enum_stats.py
@@ -8,7 +8,7 @@
 import unittest
 import tempfile
 import os
-from isoquant_lib.stats import EnumStats
+from isoquant_lib.utils.stats import EnumStats
 from enum import Enum
