+++ isoquant_tests/test_exon_splice_site_counter.py
@@ -17,7 +17,7 @@ from isoquant_lib.common import junctions_from_blocks
 from isoquant_lib.gene_info import ExonSpliceSiteRegion, GeneInfo
 from isoquant_lib.long_read_counter import ExonSpliceSiteCounter
 from isoquant_lib.isoform_assignment import MatchEventSubtype
-from isoquant_lib.string_pools import StringPoolManager
+from isoquant_lib.utils.string_pools import StringPoolManager
 
 
 # ── helpers ──────────────────────────────────────────────────────────────
