+++ isoquant_tests/test_file_naming.py
@@ -6,7 +6,7 @@
 ############################################################################
 
 import unittest
-from isoquant_lib.file_naming import *
+from isoquant_lib.utils.file_naming import *
 
 
 class TestFileNaming(unittest.TestCase):
