+++ isoquant_tests/test_file_utils.py
@@ -7,10 +7,8 @@
 ############################################################################
 
 import unittest
-import os
-import shutil
 import tempfile
-from isoquant_lib.file_utils import *
+from isoquant_lib.utils.file_utils import *
 
 
 class TestFileUtils(unittest.TestCase):
