+++ isoquant_tests/test_umi_filtering.py
@@ -12,7 +12,7 @@ from isoquant_lib.barcode_calling.umi_filtering import (
 )
 from isoquant_lib.isoform_assignment import ReadAssignment, ReadAssignmentType
 from isoquant_lib.common import junctions_from_blocks
-from isoquant_lib.string_pools import StringPoolManager
+from isoquant_lib.utils.string_pools import StringPoolManager
 
 
 class TestFormatReadAssignmentForOutput:
